Refactor: Batch 작업 레거시 코드 삭제, 패키지 구조 변경
