--- src/main/java/com/moyo/backend/domain/temporary_batch/legacy/RankingTestBatchController.java
@@ -1,22 +1,0 @@
-package com.moyo.backend.domain.temporary_batch.legacy;
-
-import lombok.RequiredArgsConstructor;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import com.moyo.backend.common.annotation.LoginUserId;
-
-@Deprecated
-@RestController
-@RequiredArgsConstructor
-public class RankingTestBatchController {
-
-	private final RankingTestBatchService rankingTestBatchService;
-
-	@GetMapping("/test/ranking/batch")
-	public void batchTest(@LoginUserId Long currentUserId) {
-
-		rankingTestBatchService.batchExec(currentUserId);
-	}
-}
