+++ src/main/java/com/moyo/backend/domain/temporary_batch/ranking/RankingBatchScheduler.java
@@ -9,8 +9,8 @@ import org.springframework.stereotype.Component;
 import com.moyo.backend.common.implement.GithubOAuthTokenReader;
 import com.moyo.backend.domain.github_ranking.implement.Ranking;
 import com.moyo.backend.domain.github_ranking.implement.RankingUpdater;
-import com.moyo.backend.domain.temporary_batch.GithubCommitStats;
-import com.moyo.backend.domain.temporary_batch.GithubRepoDetails;
+import com.moyo.backend.domain.temporary_batch.ranking.dto.GithubCommitStats;
+import com.moyo.backend.domain.temporary_batch.ranking.dto.GithubRepoDetails;
 import com.moyo.backend.domain.temporary_batch.ranking.processor.RankingCalculator;
 import com.moyo.backend.domain.temporary_batch.ranking.dto.RankingPreflight;
 import com.moyo.backend.domain.temporary_batch.ranking.dto.RankingCalculatorParameters;
