+++ src/main/java/com/moyo/backend/domain/temporary_batch/ranking/reader/GithubRankingHttpClient.java
@@ -1,9 +1,11 @@
-package com.moyo.backend.domain.temporary_batch;
+package com.moyo.backend.domain.temporary_batch.ranking.reader;
 
 import java.util.List;
 
 import org.springframework.http.ResponseEntity;
 
+import com.moyo.backend.domain.temporary_batch.ranking.dto.GithubRepoDetails;
+import com.moyo.backend.domain.temporary_batch.ranking.dto.GithubContributorDetails;
 import com.moyo.backend.domain.temporary_batch.ranking.dto.RankingPreflight;
 
 public interface GithubRankingHttpClient {
