+++ src/main/java/com/moyo/backend/domain/temporary_batch/ranking/reader/GithubRankingRestClientImpl.java
@@ -1,4 +1,4 @@
-package com.moyo.backend.domain.temporary_batch;
+package com.moyo.backend.domain.temporary_batch.ranking.reader;
 
 import static com.moyo.backend.common.constant.MoyoConstants.GITHUB_QUERY_PAGING_SIZE;
 
@@ -12,6 +12,8 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestClient;
 
+import com.moyo.backend.domain.temporary_batch.ranking.dto.GithubRepoDetails;
+import com.moyo.backend.domain.temporary_batch.ranking.dto.GithubContributorDetails;
 import com.moyo.backend.domain.temporary_batch.ranking.dto.RankingPreflight;
 
 @Slf4j
